Mas cosas de usuarios

+++ apps/user-authentication/src/auth/auth.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
 import { UsersService } from '../users/users.service';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
@@ -6,6 +6,10 @@ import { Tokens } from './types';
 import * as bcrypt from 'bcryptjs';
 import { LoginDto } from './dto/login.dto';
 import { RpcException } from '@nestjs/microservices';
+import { ChangePasswordDto } from './dto/change-password.dto';
+import { ForgotPasswordDto } from './dto/forgot-password.dto';
+import { nanoid } from 'nanoid';
+import { GetUsersFiltersDto } from '../users/dto/get-users-filter.dto';
 
 @Injectable()
 export class AuthService {
@@ -16,15 +20,9 @@ export class AuthService {
         private readonly jwtService: JwtService,
     ) {}
 
+    //Login
     async login(loginDto: LoginDto): Promise<Tokens> {      
       const user = await this.usersService.getUserByEmail(loginDto.email);
-      if (!user) {
-        throw new RpcException({
-          message: `Usuario con correo ${loginDto.email} no encontrado.`,
-          error: 'Unauthorized',
-          status: 403
-        });
-      }
   
       const passwordMatches = await bcrypt.compare(loginDto.password, user.password);
       if (!passwordMatches) {
@@ -41,16 +39,70 @@ export class AuthService {
       return tokens;
     }
   
-
+    //Logout
     async logout(userId: string){
+      if (!userId) {
+        throw new RpcException({
+            message: 'Usuario no autenticado',
+            error: 'Unauthorized',
+            status: HttpStatus.UNAUTHORIZED,
+        });
+      }
       return await this.usersService.deleteRtUser(userId)
     }
 
+    //Obtener info de usuario
+    async getInfoUser(userId: string){
+      if (!userId) {
+        throw new RpcException({
+            message: 'Usuario no autenticado',
+            error: 'Unauthorized',
+            status: HttpStatus.UNAUTHORIZED,
+        });
+      }
+      return await this.usersService.getUserByIdInfo(userId)
+    }
+
+    //Cambiar contraseña
+    async changePassword(dtoChangePassword: ChangePasswordDto, userId: string){
+      //Buscar usuario
+      const user = await this.usersService.getUserById(userId);
+
+      //Coparar contraseñas viejas
+      const oldPasswordMatches = await bcrypt.compare(dtoChangePassword.oldPassword, user.password);
+      if (!oldPasswordMatches) {
+        throw new RpcException({
+          message: 'Contraseña incorrecta. Intenta nuevamente.',
+          error: 'Unauthorized',
+          status: 403
+        });
+      }
+
+      //Encriptar contraseña nueva
+      const newHashePassword = await bcrypt.hash(dtoChangePassword.newPassword, 10)
+      return await this.usersService.changePassword(userId, newHashePassword)
+
+    }
+
+    //PENDIENTE POR EL MOMENTO ------- Cambiar contraseña pero se olvido de la antigua
+    async forgotPassword(dtoForgotPassword: ForgotPasswordDto){
+      const user = await this.usersService.getUserByEmail(dtoForgotPassword.email);
+
+      if(user){
+        const resetToken = nanoid(64)
+        await this.usersService.saveResetToken(user._id.toString(), resetToken)
+      }
+
+      return { message: "Un mensaje ha sido enviado al correo para restablece su contraseña" }
+    }
+
+    //Refresh token
     async refreshToken(userId: string, rt: string){
       const user = await this.usersService.getUserById(userId);
-      if (!user) {
+
+      if (!user.hashRefreshToken) {
         throw new RpcException({
-          message: `Usuario con id ${userId} no encontrado.`,
+          message: `Usuario no esta logeado por lo tanto no puede actualizar su token.`,
           error: 'Unauthorized',
           status: 403
         });
@@ -72,31 +124,32 @@ export class AuthService {
       
     }
 
-  async getTokens(userId: string, email: string): Promise<Tokens>{
-    const [at, rt] = await Promise.all([
+    //Obtener los tokens
+    async getTokens(userId: string, email: string): Promise<Tokens>{
+      const [at, rt] = await Promise.all([
 
-      this.jwtService.signAsync({
-        sub: userId,
-        email
-      }, {
-        secret: this.configService.getOrThrow('JWT_ACCESS_TOKEN_SECRET'),
-        expiresIn: 60 * 15,
-      }),
+        this.jwtService.signAsync({
+          sub: userId,
+          email
+        }, {
+          secret: this.configService.getOrThrow('JWT_ACCESS_TOKEN_SECRET'),
+          expiresIn: 60 * 15,
+        }),
 
-      this.jwtService.signAsync({
-        sub: userId,
-        email
-      }, {
-        secret: this.configService.getOrThrow('JWT_REFRESH_TOKEN_SECRET'),
-        expiresIn: 60 * 60 * 24 * 7,
-      })
+        this.jwtService.signAsync({
+          sub: userId,
+          email
+        }, {
+          secret: this.configService.getOrThrow('JWT_REFRESH_TOKEN_SECRET'),
+          expiresIn: 60 * 60 * 24 * 7,
+        })
 
-    ]);
+      ]);
 
-    return {
-      access_token: at,
-      refresh_token: rt,
+      return {
+        access_token: at,
+        refresh_token: rt,
+      }
     }
-  }
 
 }
